backend/requirements.txt
langchain
langchain-community
langchain-ollama
faiss-cpu
pandas
tiktoken
python-dotenv
uvicorn
fastapi
pydantic